Add useFirebase hook
fixes #149
